feat: удалять пользователей может только владелец данного рабочего пространства

+++ src/app/field/field.component.ts
@@ -2,7 +2,7 @@ import { Component, Injectable, OnInit} from '@angular/core';
 import {IListInterface} from "../interface/list.interface";
 import {ISpaceInterface} from "../interface/space.interface";
 import {WorkSpaceService} from "../shared/services/work-space.service";
-import {filter, map, switchMap} from "rxjs/operators";
+import {map, switchMap} from "rxjs/operators";
 import {UsersService} from "../shared/services/users.service";
 import {IUserInfoInterface} from "../interface/user-info.interface";
 import {Observable} from "rxjs";
@@ -95,17 +95,20 @@ export class FieldComponent implements OnInit {
   }
 
   handleDeletedWorkspaceOwner($event: IUserInfoInterface) {
-    console.log('$event',$event)
-    this.workSpaceService.deleteUserWorkSpace($event._id,  this.currentSpace._id)
-      .subscribe(
-      (v: ISpaceInterface) => {
-        this.currentSpace.owner = this.currentSpace?.owner.filter((val) => val !== $event._id)
-        this.usersWorkSpaceOwner$ = this.usersWorkSpaceOwner$.pipe(map((value: IUserInfoInterface[]) => {
-         return value.filter((item: IUserInfoInterface) => item._id !== $event._id)
-        }))
 
-      }
-    )
-
+    if(JSON.parse(localStorage.getItem('userInfo'))._id === this.currentSpace?.holder) {
+      this.workSpaceService.deleteUserWorkSpace($event._id,  this.currentSpace._id)
+        .subscribe(
+          () => {
+            this.currentSpace.owner = this.currentSpace?.owner.filter((val) => val !== $event._id)
+            this.usersWorkSpaceOwner$ = this.usersWorkSpaceOwner$.pipe(map((value: IUserInfoInterface[]) => {
+              return value.filter((item: IUserInfoInterface) => item._id !== $event._id)
+            }))
+          }
+        )
+    } else {
+      console.log('Нельзя удалять пользователей если вы не владелец данного рабочего пространства');
+    }
   }
+
 }
